[CI] Split Autoscaler e2e tests into 2 buildkite runners (#3715)
